+++ src/components/Input/PasswordInput.stories.tsx
@@ -1,5 +1,5 @@
 import { StoryFn } from '@storybook/react';
-import PasswordInput from './PasswordInput';
+import { PasswordInput } from './PasswordInput';
 
 export default {
   title: 'PasswordInput',
