+++ src/components/Input/SecurityCodeInput.stories.tsx
@@ -1,5 +1,5 @@
 import { StoryFn } from '@storybook/react';
-import SecurityCodeInput from './SecurityCodeInput';
+import { SecurityCodeInput } from './SecurityCodeInput';
 
 export default {
   title: 'SecurityCodeInput',
