+++ src/components/Input/index.js
@@ -1,0 +1,5 @@
+export * from './CardNumberInput';
+export * from './ExpirationDateInput';
+export * from './NameInput';
+export * from './PasswordInput';
+export * from './SecurityCodeInput';
