+++ src/components/common/Label/Label.tsx
@@ -1,11 +1,9 @@
 export type Props = { text: string };
 
-function Label({ text }: Props) {
+export function Label({ text }: Props) {
   return (
     <>
       <span>{text}</span>
     </>
   );
 }
-
-export default Label;
