+++ src/components/common/index.js
@@ -1,0 +1,2 @@
+export * from './Label/Label';
+export * from './Input/Input';
